move ticket tasks to own file
